@@ -2,7 +2,7 @@
   "name": "kilogram",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "index.ts",
   "dependencies": {
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
@@ -10,8 +10,8 @@
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^13.1.5",
-    "cors": "^2.8.5",
     "concurrently": "^6.0.2",
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "express-session": "^1.17.2",
     "react": "^17.0.2",
@@ -19,7 +19,8 @@
     "react-redux": "^7.2.3",
     "react-scripts": "^4.0.3",
     "sass": "^1.32.10",
-    "typescript": "^4.2.4"
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.4"
   },
   "repository": {
     "type": "git",
@@ -55,6 +56,7 @@
   },
   "license": "ISС",
   "devDependencies": {
+    "@types/express": "^4.17.12",
     "@typescript-eslint/eslint-plugin": "^4.21.0",
     "@typescript-eslint/parser": "^4.21.0",
     "dotenv": "^8.2.0",
