@@ -1,3 +1,4 @@
-module.exports = {
+export const config = {
+	host: "https://kilogram-team-2.herokuapp.com",
 	serverPort: 9000,
 }
