@@ -1,20 +1,23 @@
 import {Avatar} from "@material-ui/core"
 import React from "react"
 
-const Message = () => {
+const message = (login) => {
 	return (
 		<div className="Message">
-			<Avatar style={{alignSelf: "flex-end"}} />
+			<Avatar
+				style={{alignSelf: "flex-end"}}
+				src="https://thiscatdoesnotexist.com/"
+			/>
 			<div className="MessageBody">
 				<p>
-					Hi! Lorem ipsum dolor sit amet consectetur adipisicing elit.
-					Quidem saepe suscipit dolores aspernatur, explicabo atque
-					sed nisi repellendus numquam, ea alias sit quas libero
-					tempore culpa quia id ab sunt!
+          Hi, {login}! Lorem ipsum dolor sit amet consectetur adipisicing elit.
+          Quidem saepe suscipit dolores aspernatur, explicabo atque sed nisi
+          repellendus numquam, ea alias sit quas libero tempore culpa quia id ab
+          sunt!
 				</p>
 			</div>
 		</div>
 	)
 }
 
-export default Message
+export default message
